+++ include/ice/manage/dec/StreamingDecoder.hpp
@@ -17,8 +17,8 @@ class StreamingDecoder : public IDecoder {
         std::shared_ptr<IDecoderFactory> factory);
 
     // 解码数据到缓冲区的接口
-    size_t decode(float** buffer, uint16_t num_channels, size_t start_frame,
-                  size_t frame_count) override;
+    double decode(float** buffer, uint16_t num_channels, double start_frame,
+                  double frame_count) override;
 
     size_t num_frames() const override { return 0; }
